Start refactor to inverse
